Refactor tests and schema fields for system and user tables
Updated test setup to use init_test_app_context for consistent initialization across modules. Adjusted system tables and user table schema tests to expect 8 core fields instead of 11, reflecting recent schema changes. Improved UPDATE statement parsing for table references. Changed test column type from Int to BigInt in user table provider tests.

+++ backend/crates/kalamdb-core/src/tables/user_tables/user_table_provider.rs
@@ -741,7 +741,7 @@ mod tests {
             ColumnDefinition::new(
                 "id",
                 1,
-                KalamDataType::Int,
+                KalamDataType::BigInt,  // Use BigInt for Int64
                 false, // not nullable
                 true,  // primary key
                 false, // not unique
